requirements.txt
alembic==1.16.1
anndata==0.10.9
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings @ file:///croot/argon2-cffi-bindings_1736182440035/work
array_api_compat==1.8
arrow==1.3.0
asttokens @ file:///croot/asttokens_1743630435401/work
async-lru @ file:///work/perseverance-python-buildout/croot/async-lru_1701732681408/work
attrs==25.3.0
babel==2.17.0
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1718029820055/work
black==25.1.0
bleach @ file:///croot/bleach_1732290411627/work
blinker==1.9.0
Brotli @ file:///croot/brotli-split_1736182456865/work
cachetools==5.5.2
certifi @ file:///croot/certifi_1745939216646/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
comm @ file:///croot/comm_1709322850197/work
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.54.0
debugpy @ file:///croot/debugpy_1736267418885/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
docker==7.1.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema @ file:///croot/python-fastjsonschema_1731939362158/work
filelock==3.18.0
flake8==7.2.0
Flask==3.1.1
fonttools==4.58.0
fqdn==1.5.1
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.2
gunicorn==23.0.0
h11 @ file:///croot/h11_1748442006460/work
h5py==3.11.0
hf-xet==1.1.3
httpcore @ file:///croot/httpcore_1748526048470/work
httpx @ file:///croot/httpx_1746747840559/work
huggingface-hub==0.32.4
identify==2.6.10
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
ipykernel @ file:///croot/ipykernel_1737660677549/work
ipython @ file:///croot/ipython_1744912971218/work
ipython_pygments_lexers @ file:///croot/ipython_pygments_lexers_1744753235686/work
ipywidgets @ file:///croot/ipywidgets_1733504575802/work
isoduration==20.11.0
isort==6.0.1
itsdangerous==2.2.0
jedi @ file:///croot/jedi_1733987392413/work
Jinja2 @ file:///croot/jinja2_1741710844255/work
jiter==0.10.0
joblib==1.5.1
json5 @ file:///croot/json5_1730786798687/work
jsonpointer==3.0.0
jsonschema @ file:///croot/jsonschema_1728486696720/work
jsonschema-specifications==2025.4.1
jupyter @ file:///croot/jupyter_1737645803214/work
jupyter-console @ file:///work/perseverance-python-buildout/croot/jupyter_console_1707343324117/work
jupyter-events @ file:///croot/jupyter_events_1741184577592/work
jupyter-lsp @ file:///croot/jupyter-lsp-meta_1745827015155/work
jupyter_client @ file:///croot/jupyter_client_1737570961872/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
jupyter_server @ file:///croot/jupyter_server_1741200034823/work
jupyter_server_terminals @ file:///croot/jupyter_server_terminals_1744706698694/work
jupyterlab==4.4.3
jupyterlab_pygments @ file:///croot/jupyterlab_pygments_1741124142640/work
jupyterlab_server @ file:///croot/jupyterlab_server_1725865349919/work
jupyterlab_widgets @ file:///croot/jupyterlab_widgets_1733440870459/work
kiwisolver==1.4.8
Levenshtein==0.27.1
llvmlite==0.43.0
Mako==1.3.10
Markdown==3.8
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.3
matplotlib-inline @ file:///work/perseverance-python-buildout/croot/matplotlib-inline_1698864771271/work
mccabe==0.7.0
mistune @ file:///croot/mistune_1741124011532/work
mlflow==2.22.0
mlflow-skinny==2.22.0
mpmath==1.3.0
mypy==1.15.0
mypy_extensions==1.1.0
natsort==8.4.0
nbclient @ file:///croot/nbclient_1741123995822/work
nbconvert @ file:///croot/nbconvert-meta_1741184653315/work
nbformat @ file:///croot/nbformat_1728049424075/work
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.3
nodeenv==1.9.1
notebook==7.4.3
notebook_shim @ file:///croot/notebook-shim_1741707758683/work
numba==0.60.0
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.4.8
openai==1.82.0
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
overrides @ file:///work/perseverance-python-buildout/croot/overrides_1701732220415/work
packaging @ file:///croot/packaging_1734472117206/work
pandas==2.2.3
pandocfilters==1.5.1
parso @ file:///croot/parso_1733963305961/work
pathspec==0.12.1
patsy==0.5.6
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
prometheus_client @ file:///croot/prometheus_client_1744271615306/work
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==6.31.0
psutil==7.0.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments @ file:///croot/pygments_1744664109463/work
pynndescent==0.5.13
pyparsing==3.2.3
PyQt6==6.7.1
PyQt6_sip @ file:///croot/pyqt-split_1744804475988/work/pyqt_sip
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq @ file:///croot/pyzmq_1734687138743/work
qtconsole @ file:///croot/qtconsole_1744633877544/work
QtPy @ file:///work/perseverance-python-buildout/croot/qtpy_1701733558435/work
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
rfc3339-validator @ file:///work/perseverance-python-buildout/croot/rfc3339-validator_1698873973408/work
rfc3986-validator @ file:///work/perseverance-python-buildout/croot/rfc3986-validator_1698874010100/work
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scikit-misc==0.5.1
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
session-info==1.0.0
setuptools==80.8.0
sip @ file:///croot/sip_1738856193618/work
six @ file:///croot/six_1744271502820/work
smmap==5.0.2
sniffio==1.3.1
soupsieve @ file:///work/perseverance-python-buildout/croot/soupsieve_1698866207280/work
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.46.2
statsmodels==0.14.2
stdlib-list==0.10.0
sympy==1.13.1
terminado @ file:///work/perseverance-python-buildout/croot/terminado_1698874934826/work
threadpoolctl==3.6.0
tinycss2 @ file:///croot/tinycss2_1738337643607/work
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.5.1
tqdm==4.67.1
traitlets @ file:///croot/traitlets_1718227057033/work
transformers==4.52.4
triton==3.2.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.6
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.31.2
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==24.11.1
webencodings @ file:///work/perseverance-python-buildout/croot/webencodings_1698866454420/work
websocket-client @ file:///croot/websocket-client_1715878298792/work
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension @ file:///croot/widgetsnbextension_1733439572884/work
wrapt==1.17.2
zipp==3.21.0
